Updated cloud functions logic to automate the workflow
