--- cloud_functions/climateiq_aggregate_clipping/requirements.txt
@@ -1,3 +1,0 @@
-rasterio
-geopandas
-google-cloud-storage
