+++ cloud_functions/climateiq_process_rsl_cf/requirements.txt
@@ -1,2 +1,3 @@
-pandas
-google-cloud-storage
+google-cloud-storage==2.10.0
+pandas==1.4.4
+functions-framework==3.2.0
