--- cloud_functions/climateiq_tiff_to_h3/requirements.txt
@@ -1,6 +1,0 @@
-pandas
-google-cloud-storage
-rasterio
-numpy
-h3
-shapely
