+++ cloud_functions/climateiq_tiff_to_h3_cf/requirements.txt
@@ -1,0 +1,6 @@
+google-cloud-storage==2.5.0
+h3==3.7.1
+rasterio==1.2.10
+pandas==1.3.3
+shapely==1.7.1
+numpy==1.21.2
